🖥️ wip: refactored: movie--detail page
