+++ src/components/Movie/index.ts
@@ -1,2 +1,4 @@
-export * from './Movies';
-export { default } from './Movies';
+export { default as MovieCast } from './MovieCast';
+export { default as MovieGallery } from './MovieGallery';
+export { default as MovieTorrents } from './MovieTorrents';
+export { default as useMovieStyles } from './Movies.style';
